added/edited migration user_competencies. edited view competentie kiezen. added some non working code snippets - will clean up later - should be commented right now, probably
