+++ app/Http/Controllers/UserCompetenciesController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class UserCompetenciesController extends Controller
+{
+    //
+}
