+++ resources/views/project/index.blade.php
@@ -47,7 +47,7 @@
 								{!! Form::close() !!}
 						</div>
 					</td>
-					
+
 				</tr>
 				@endforeach
 			</tbody>
